backend/requirements.txt
wheel==0.41.3
flake8==6.1.0
Django==4.2.7
# sterowniki do postgresa pod django
psycopg2==2.9.9
djangorestframework==3.14.0
